WIP: Begin extracting `ReplStore`
