+++ crates/repl/src/repl.rs
@@ -1,11 +1,13 @@
 use async_dispatcher::{set_dispatcher, Dispatcher, Runnable};
 use gpui::{AppContext, PlatformDispatcher};
+use project::Fs;
 use settings::Settings as _;
 use std::{sync::Arc, time::Duration};
 
 mod jupyter_settings;
 mod kernels;
 mod outputs;
+mod repl_store;
 mod runtime_panel;
 mod session;
 mod stdio;
@@ -16,6 +18,8 @@ pub use runtime_panel::{ClearOutputs, Interrupt, Run, Shutdown};
 pub use runtime_panel::{RuntimePanel, SessionSupport};
 pub use runtimelib::ExecutionState;
 pub use session::Session;
+
+use crate::repl_store::ReplStore;
 
 fn zed_dispatcher(cx: &mut AppContext) -> impl Dispatcher {
     struct ZedDispatcher {
@@ -41,8 +45,9 @@ fn zed_dispatcher(cx: &mut AppContext) -> impl Dispatcher {
     }
 }
 
-pub fn init(cx: &mut AppContext) {
+pub fn init(fs: Arc<dyn Fs>, cx: &mut AppContext) {
     set_dispatcher(zed_dispatcher(cx));
     JupyterSettings::register(cx);
-    runtime_panel::init(cx)
+    runtime_panel::init(cx);
+    ReplStore::init(fs, cx);
 }
