+++ crates/repl/src/runtime_panel.rs
@@ -1,3 +1,4 @@
+use crate::repl_store::ReplStore;
 use crate::{
     jupyter_settings::{JupyterDockPosition, JupyterSettings},
     kernels::{kernel_specifications, KernelSpecification},
@@ -8,7 +9,7 @@ use collections::HashMap;
 use editor::{Anchor, Editor, RangeToAnchorExt};
 use gpui::{
     actions, prelude::*, AppContext, AsyncWindowContext, EntityId, EventEmitter, FocusHandle,
-    FocusOutEvent, FocusableView, Subscription, Task, View, WeakView,
+    FocusOutEvent, FocusableView, Subscription, Task, View, WeakView, WindowContext,
 };
 use language::{Language, Point};
 use multi_buffer::MultiBufferRow;
@@ -36,11 +37,10 @@ pub fn init(cx: &mut AppContext) {
             });
 
             workspace.register_action(|workspace, _: &RefreshKernelspecs, cx| {
-                if let Some(panel) = workspace.panel::<RuntimePanel>(cx) {
-                    panel.update(cx, |panel, cx| {
-                        panel.refresh_kernelspecs(cx).detach();
-                    });
-                }
+                let store = ReplStore::global(cx);
+                store.update(cx, |store, cx| {
+                    store.refresh_kernelspecs(cx).detach();
+                });
             });
         },
     )
@@ -67,7 +67,10 @@ pub fn init(cx: &mut AppContext) {
                     let weak_editor = cx.view().downgrade();
                     panel.update(cx, |_, cx| {
                         cx.defer(|panel, cx| {
-                            panel.run(weak_editor, cx).log_err();
+                            let store = ReplStore::global(cx);
+                            store.update(cx, |store, cx| {
+                                store.run(weak_editor, cx).log_err();
+                            });
                         });
                     })
                 },
@@ -145,11 +148,8 @@ pub fn init(cx: &mut AppContext) {
 
 pub struct RuntimePanel {
     fs: Arc<dyn Fs>,
-    enabled: bool,
     focus_handle: FocusHandle,
     width: Option<Pixels>,
-    sessions: HashMap<EntityId, View<Session>>,
-    kernel_specifications: Vec<KernelSpecification>,
     _subscriptions: Vec<Subscription>,
 }
 
@@ -168,19 +168,13 @@ impl RuntimePanel {
                     let subscriptions = vec![
                         cx.on_focus_in(&focus_handle, Self::focus_in),
                         cx.on_focus_out(&focus_handle, Self::focus_out),
-                        cx.observe_global::<SettingsStore>(move |this, cx| {
-                            this.set_enabled(JupyterSettings::enabled(cx), cx);
-                        }),
                     ];
 
                     let runtime_panel = Self {
-                        fs: fs.clone(),
+                        fs,
                         width: None,
                         focus_handle,
-                        kernel_specifications: Vec::new(),
-                        sessions: Default::default(),
                         _subscriptions: subscriptions,
-                        enabled: JupyterSettings::enabled(cx),
                     };
 
                     runtime_panel
@@ -194,228 +188,12 @@ impl RuntimePanel {
         })
     }
 
-    fn set_enabled(&mut self, enabled: bool, cx: &mut ViewContext<Self>) {
-        if self.enabled != enabled {
-            self.enabled = enabled;
-            cx.notify();
-        }
-    }
-
     fn focus_in(&mut self, cx: &mut ViewContext<Self>) {
         cx.notify();
     }
 
     fn focus_out(&mut self, _event: FocusOutEvent, cx: &mut ViewContext<Self>) {
         cx.notify();
-    }
-
-    pub fn snippet(
-        &self,
-        editor: WeakView<Editor>,
-        cx: &mut ViewContext<Self>,
-    ) -> Option<(String, Arc<Language>, Range<Anchor>)> {
-        let editor = editor.upgrade()?;
-        let editor = editor.read(cx);
-
-        let buffer = editor.buffer().read(cx).snapshot(cx);
-
-        let selection = editor.selections.newest::<usize>(cx);
-        let multi_buffer_snapshot = editor.buffer().read(cx).snapshot(cx);
-
-        let range = if selection.is_empty() {
-            let cursor = selection.head();
-
-            let cursor_row = multi_buffer_snapshot.offset_to_point(cursor).row;
-            let start_offset = multi_buffer_snapshot.point_to_offset(Point::new(cursor_row, 0));
-
-            let end_point = Point::new(
-                cursor_row,
-                multi_buffer_snapshot.line_len(MultiBufferRow(cursor_row)),
-            );
-            let end_offset = start_offset.saturating_add(end_point.column as usize);
-
-            // Create a range from the start to the end of the line
-            start_offset..end_offset
-        } else {
-            selection.range()
-        };
-
-        let anchor_range = range.to_anchors(&multi_buffer_snapshot);
-
-        let selected_text = buffer
-            .text_for_range(anchor_range.clone())
-            .collect::<String>();
-
-        let start_language = buffer.language_at(anchor_range.start)?;
-        let end_language = buffer.language_at(anchor_range.end)?;
-        if start_language != end_language {
-            return None;
-        }
-
-        Some((selected_text, start_language.clone(), anchor_range))
-    }
-
-    pub fn language(
-        &self,
-        editor: WeakView<Editor>,
-        cx: &mut ViewContext<Self>,
-    ) -> Option<Arc<Language>> {
-        let editor = editor.upgrade()?;
-        let selection = editor.read(cx).selections.newest::<usize>(cx);
-        let buffer = editor.read(cx).buffer().read(cx).snapshot(cx);
-        buffer.language_at(selection.head()).cloned()
-    }
-
-    pub fn refresh_kernelspecs(&mut self, cx: &mut ViewContext<Self>) -> Task<anyhow::Result<()>> {
-        let kernel_specifications = kernel_specifications(self.fs.clone());
-        cx.spawn(|this, mut cx| async move {
-            let kernel_specifications = kernel_specifications.await?;
-
-            this.update(&mut cx, |this, cx| {
-                this.kernel_specifications = kernel_specifications;
-                cx.notify();
-            })
-        })
-    }
-
-    pub fn kernelspec(
-        &self,
-        language: &Language,
-        cx: &mut ViewContext<Self>,
-    ) -> Option<KernelSpecification> {
-        let settings = JupyterSettings::get_global(cx);
-        let language_name = language.code_fence_block_name();
-        let selected_kernel = settings.kernel_selections.get(language_name.as_ref());
-
-        self.kernel_specifications
-            .iter()
-            .find(|runtime_specification| {
-                if let Some(selected) = selected_kernel {
-                    // Top priority is the selected kernel
-                    runtime_specification.name.to_lowercase() == selected.to_lowercase()
-                } else {
-                    // Otherwise, we'll try to find a kernel that matches the language
-                    runtime_specification.kernelspec.language.to_lowercase()
-                        == language_name.to_lowercase()
-                }
-            })
-            .cloned()
-    }
-
-    pub fn run(
-        &mut self,
-        editor: WeakView<Editor>,
-        cx: &mut ViewContext<Self>,
-    ) -> anyhow::Result<()> {
-        if !self.enabled {
-            return Ok(());
-        }
-
-        let (selected_text, language, anchor_range) = match self.snippet(editor.clone(), cx) {
-            Some(snippet) => snippet,
-            None => return Ok(()),
-        };
-
-        let entity_id = editor.entity_id();
-
-        let kernel_specification = self
-            .kernelspec(&language, cx)
-            .with_context(|| format!("No kernel found for language: {}", language.name()))?;
-
-        let session = self.sessions.entry(entity_id).or_insert_with(|| {
-            let view =
-                cx.new_view(|cx| Session::new(editor, self.fs.clone(), kernel_specification, cx));
-            cx.notify();
-
-            let subscription = cx.subscribe(
-                &view,
-                |panel: &mut RuntimePanel, _session: View<Session>, event: &SessionEvent, _cx| {
-                    match event {
-                        SessionEvent::Shutdown(shutdown_event) => {
-                            panel.sessions.remove(&shutdown_event.entity_id());
-                        }
-                    }
-                },
-            );
-
-            subscription.detach();
-
-            view
-        });
-
-        session.update(cx, |session, cx| {
-            session.execute(&selected_text, anchor_range, cx);
-        });
-
-        anyhow::Ok(())
-    }
-
-    pub fn clear_outputs(&mut self, editor: WeakView<Editor>, cx: &mut ViewContext<Self>) {
-        let entity_id = editor.entity_id();
-        if let Some(session) = self.sessions.get_mut(&entity_id) {
-            session.update(cx, |session, cx| {
-                session.clear_outputs(cx);
-            });
-            cx.notify();
-        }
-    }
-
-    pub fn interrupt(&mut self, editor: WeakView<Editor>, cx: &mut ViewContext<Self>) {
-        let entity_id = editor.entity_id();
-        if let Some(session) = self.sessions.get_mut(&entity_id) {
-            session.update(cx, |session, cx| {
-                session.interrupt(cx);
-            });
-            cx.notify();
-        }
-    }
-
-    pub fn shutdown(&self, editor: WeakView<Editor>, cx: &mut ViewContext<RuntimePanel>) {
-        let entity_id = editor.entity_id();
-        if let Some(session) = self.sessions.get(&entity_id) {
-            session.update(cx, |session, cx| {
-                session.shutdown(cx);
-            });
-            cx.notify();
-        }
-    }
-}
-
-pub enum SessionSupport {
-    ActiveSession(View<Session>),
-    Inactive(Box<KernelSpecification>),
-    RequiresSetup(Arc<str>),
-    Unsupported,
-}
-
-impl RuntimePanel {
-    pub fn session(
-        &mut self,
-        editor: WeakView<Editor>,
-        cx: &mut ViewContext<Self>,
-    ) -> SessionSupport {
-        let entity_id = editor.entity_id();
-        let session = self.sessions.get(&entity_id).cloned();
-
-        match session {
-            Some(session) => SessionSupport::ActiveSession(session),
-            None => {
-                let language = self.language(editor, cx);
-                let language = match language {
-                    Some(language) => language,
-                    None => return SessionSupport::Unsupported,
-                };
-                let kernelspec = self.kernelspec(&language, cx);
-
-                match kernelspec {
-                    Some(kernelspec) => SessionSupport::Inactive(Box::new(kernelspec)),
-                    None => match language.name().as_ref() {
-                        "TypeScript" | "Python" => SessionSupport::RequiresSetup(language.name()),
-                        _ => SessionSupport::Unsupported,
-                    },
-                }
-            }
-        }
     }
 }
 
@@ -461,8 +239,10 @@ impl Panel for RuntimePanel {
         self.width = size;
     }
 
-    fn icon(&self, _cx: &ui::WindowContext) -> Option<ui::IconName> {
-        if !self.enabled {
+    fn icon(&self, cx: &ui::WindowContext) -> Option<ui::IconName> {
+        let store = ReplStore::global(cx);
+
+        if !store.read(cx).is_enabled() {
             return None;
         }
 
@@ -488,10 +268,19 @@ impl FocusableView for RuntimePanel {
 
 impl Render for RuntimePanel {
     fn render(&mut self, cx: &mut ViewContext<Self>) -> impl IntoElement {
+        let store = ReplStore::global(cx);
+
+        let (kernel_specifications, sessions) = store.update(cx, |store, cx| {
+            (
+                store.kernel_specifications().cloned().collect::<Vec<_>>(),
+                store.sessions().cloned().collect::<Vec<_>>(),
+            )
+        });
+
         // When there are no kernel specifications, show a link to the Zed docs explaining how to
         // install kernels. It can be assumed they don't have a running kernel if we have no
         // specifications.
-        if self.kernel_specifications.is_empty() {
+        if kernel_specifications.is_empty() {
             return v_flex()
                 .p_4()
                 .size_full()
@@ -519,7 +308,7 @@ impl Render for RuntimePanel {
         }
 
         // When there are no sessions, show the command to run code in an editor
-        if self.sessions.is_empty() {
+        if sessions.is_empty() {
             return v_flex()
                 .p_4()
                 .size_full()
@@ -539,7 +328,7 @@ impl Render for RuntimePanel {
                 )
                 .child(Label::new("Kernels available").size(LabelSize::Large))
                 .children(
-                    self.kernel_specifications.iter().map(|spec| {
+                    kernel_specifications.into_iter().map(|spec| {
                         h_flex().gap_2().child(Label::new(spec.name.clone()))
                             .child(Label::new(spec.kernelspec.language.clone()).color(Color::Muted))
                     })
@@ -552,8 +341,8 @@ impl Render for RuntimePanel {
             .p_4()
             .child(Label::new("Jupyter Kernel Sessions").size(LabelSize::Large))
             .children(
-                self.sessions
-                    .values()
+                sessions
+                    .into_iter()
                     .map(|session| session.clone().into_any_element()),
             )
             .into_any_element()
